Get tests mostly working; fix "import server" issue
