+++ src/typescript/.env.ci
@@ -9,9 +9,9 @@ NEXT_PUBLIC_MODULE_ADDRESS="0xf000000d12c21b4170c7165010a4373b9ebc47bb9be9f0e480
 # The address of the rewards module with the overloaded swap function
 NEXT_PUBLIC_REWARDS_MODULE_ADDRESS="0xf000000d12c21b4170c7165010a4373b9ebc47bb9be9f0e48091645308a14136"
 
-# The default URL for a local inbox deployment. Do not add a slash to the end
-# Set through CI env variables
-INBOX_URL=""
+# The default URL for a local indexer deployment. Do not add a slash to the end.
+# Set through CI env variables.
+INDEXER_URL="http://localhost:3000"
 
 ## A `nextjs` setting to determine the query revalidation length, in seconds.
 REVALIDATION_TIME="1"
