+++ src/typescript/sdk/src/queries/inbox-url.ts
@@ -3,7 +3,7 @@ import "server-only";
 import { PostgrestClient } from "@supabase/postgrest-js";
 
 if (!process.env.INBOX_URL) {
-  throw new Error("The `inbox` endpoint is not defined for server actions.");
+  console.warn("Remove this file upon submission of this PR.")
 }
 
 const authorizationHeaders = process.env.INBOX_JWT_TOKEN
