+++ src/typescript/sdk/tests/utils/aptos-client.ts
@@ -42,7 +42,7 @@ export async function fundAccountFast(
   maybeAddress: AccountAddressInput | Account,
   // Note that this is the display amount; i.e., 1 => 10^8 input amount on-chain.
   aptAmount: number | bigint
-): Promise<void> {
+) {
   const isHexInput = typeof maybeAddress === "string" || maybeAddress instanceof Uint8Array;
   const accountAddress = AccountAddress.from(
     maybeAddress instanceof AccountAddress || isHexInput
@@ -50,11 +50,11 @@ export async function fundAccountFast(
       : maybeAddress.accountAddress
   );
   if (getAptosNetwork() === Network.LOCAL) {
-    await aptos.fundAccount({ accountAddress, amount: Number(aptAmount) * ONE_APT });
+    return [await aptos.fundAccount({ accountAddress, amount: Number(aptAmount) * ONE_APT })];
   } else {
     const results = Array.from({ length: Number(aptAmount) }).map(async () =>
       aptos.fundAccount({ accountAddress, amount: ONE_APT })
     );
-    await Promise.all(results);
+    return await Promise.all(results);
   }
 }
